Move (mostly deprecated) frontend detection to io.formats.console
